When encoding record elements, check if indentation is required
Nested record element values that do not immediately begin with a
newline character must have a space between the colon (after the key)
and the encoded value.

E.g. YAML interprets this as a string:

```
test:[]
```

But this as a record containing a single key ("test") with a value
being an empty list:

```
test: []
```